requirements.txt
# Discord Bot Dependencies
discord.py>=2.7.0
python-dotenv==1.0.0
aiosqlite==0.19.0
audioop-lts>=0.2.0
